French car market data

This data set shows monthly data of the French car market between 2003 and 2015. The market is divided into 5 main segments (A to E), according to the size of the vehicle chassis. Morais et. al (2018) first used this data to compare compositional and Dirichlet models for market shares.

Usage

car_market

Format

An object of class data.frame with 152 rows and 10 columns.

Source

The figures for GDP and household expenditure are originally provided by the The National Institute of Statistics and Economic Studies (INSEE).

The gas prices are from the OECD.

The market share of each segment of come from a simulation by Renault.

Details

SEG_: Corresponds to the shares of sales in each of the five market segments A,B,C,D and E. Where A are the smallest cars and E the largest. The segmentation is explained in wikipedia.

GDP: GDP figures in millions at current prices

HOUSEHOLD_EXPENDITURE: total household expenditure in millions at previous years prices

GAS_PRICE: Corresponds to the gas price including VAT.

SCRAPPING_SUBSIDY: A dummy indicating periods where the French government provided subsidies for scrapping a car.

References

Joanna Morais, Christine Thomas-Agnan & Michel Simioni (2018) Using compositional and Dirichlet models for market share regression, Journal of Applied Statistics, 45:9, 1670-1689, DOI: 10.1080/02664763.2017.1389864

Author

Lukas Dargel, Christine Thomas-Agnan