Data on the rice yields in the Vietnamese provinces

The data is used in the article by (Huong et al. 2023), who study the impact of climate change on the rice production in Vietnam.
It contains the following information:

PROVINCE: a factor for the 63 provinces of Vietnam

REGION: a factor with the 6 main regions

YEAR: a numeric corresponding to the year

YIELD: a numeric for the rice production in tonnes per hectare

PRECIPITATION: a numeric for the annual precipitation in liters

TEMPERATURES: a compositional variable represented as a matrix
It corresponding to the fraction of days in a year where the maximal temperature falls into one of the three categories, "LOW", "MIDDLE" and "HIGH".

Usage

rice_yields

Format

An object of class data.frame with 1890 rows and 6 columns.

References

Thi-Huong Trinh, Michel Simioni, and Christine Thomas-Agnan, “Discrete and Smooth Scalar-on-Density Compositional Regression for Assessing the Impact of Climate Change on Rice Yield in Vietnam”, TSE Working Paper, n. 23-1410, February 2023.

Author

Lukas Dargel, Christine Thomas-Agnan