Simulated retail data for nine shopping malls in the city of Toulouse

This data set provides an example for the use of CoDa models in geomarketing applications. The data is simulated, but realistic in the sense that the parameters used for the simulation were estimated on a real, but confidential data set (Dargel and Thomas-Agnan 2024).

Usage

toulouse_retail

Format

An object of class sf (inherits from data.table, data.frame) with 428 rows and 6 columns.

Source

The figures for POP and MEDIAN_INCOME come from the French census data prided by INSEE.

The polygon geometry is provided by the IGN.

The locations of the nine shopping malls around the city center are derived from online mapping services (Google Maps and OpenStreetMap).

The distances (dist_km) are derived from location information.

The number of shopping trips (visits) are simulated by the authors.

Details

ID_IRIS: Identifies the geographic region.

POP: Population number within the region

MEDIAN_INCOME: The median income in the region

dist_km: Distances from the region to all nine shopping malls.

visits: The share of visitors that coming from the region and going to each of the malls.

geometry: The geometry (polygon) of the region

The "mall_locations" and the "simulation_parameters" are given as additional attributes.

References

Lukas Dargel & Christine Thomas-Agnan (2024) “The link between multiplicative competitive interaction models and compositional data regression with a total”, Journal of Applied Statistics, DOI: 10.1080/02664763.2024.2329923

Author

Lukas Dargel, Christine Thomas-Agnan